refactor(admin): `Professor` 디렉토리 하위 컴포넌트 이동
