+++ apps/admin/src/features/professor/components/ProfessorCreator.tsx
@@ -5,7 +5,7 @@ import type { ProfessorRequest } from '~/apis/admin/requests'
 import { useProfessorServiceGetApiV1ProfessorsKey } from '~/apis/community/queries'
 
 import { useModal } from '~/hooks/use-modal'
-import { queryClient } from '~/utils/get-query-client'
+import { queryClient } from '~/shared/utils'
 
 function CreateProfessorForm({ onClose }: { onClose: () => void }) {
   const [form] = Form.useForm()
