+++ apps/admin/src/features/professor/components/ProfessorTable.tsx
@@ -10,7 +10,7 @@ import { useProfessorServiceGetApiV1ProfessorsKey } from '~/apis/community/queri
 import { useProfessorServiceGetApiV1ProfessorsSuspense } from '~/apis/community/queries/suspense'
 import type { ProfessorResponse } from '~/apis/community/requests'
 import useEditTable from '~/hooks/use-edit-table'
-import ProfessorTableView from './professor-table-view'
+import ProfessorTableView from './ProfessorTableView'
 
 function ProfessorTable() {
   const [form] = Form.useForm()
